Implemented material icons package
This changes most inline svg icons into the material icons svg npm package.

This is a potential breaking change from consumers of this library as you will either end up with an inline svg or a base64 data-url depending on how this is consumed which may affect how the css works out. In theory because we are in shadow-dom, it should not be an issue but it is worth testing before upgrading to this new version.

For stenciljs consumers of this package you can simply
`npm install stencil-inline-svg --save-dev` and call it in your stencil config plugins, for usage in other frameworks, please open an issue with minimal usage scenario for help.