Resolved multiple warnings about changing state during render
A warning shows whenever a state that triggers a re-render happens as it is usually something that should be done before rendering and an extra render can be avoided by moving it to another lifecycle event. However in some cases it does make sense to re-render on purpose. The way to avoid the warning is to wrap the method in a requestAnimationFrame promise so that the state changes when we are sure the initial rendering is totally complete. It does not change much as it still re-renders but it kinda declares that it was on purpose and we know what we are doing. This way the warning goes away.