Created story for collapsible
This adds a complete story for dnn-collapsible including actions hooks and supported properties.

Also this brings in lit-html which is not LitElements, it is only a very small independent par of it that allows almost jsx support which allows to dynamically create html from custom elements. Under the hood it creates an html template and keeps a reference to all dynamic parts, then injects a clone of that component with the dynamic parts replaced each time the function get's called.

Soooo, I initially though that the WebComponents docs where made only for LitElements but in fact it just used one independed function from them. See https://lit.dev/docs/v1/lit-html/introduction/ for some intro and https://lit.dev/docs/v1/lit-html/template-reference/ for some more details on the binding types. Basically `?` prefix for bolean attributes, `.` prefix for properties that are not attributes and `@` prefix for event listeners. We may just need `?` in our use cases though. There is also a vscode extension called lit-html which does syntax highlighting almost like in jsx/tsx.

One more note, I am not totally set on the idea, so we can chat about it, but if the last part of the path matches exactly the only named export (there is a single story), then the component in the list is hoisted up one level so it saves the user to have to expand that level on the left list. But apparently it needs to be a single word, soooo, we can play with variations maybe there is some science to this auto-magic.