Upgraded Storybook to v6.5.12
Storybook was broken on github pages after the last round of dependency updates. Trying to fix this locally did not prove very useful as it was not broken locally, but I found they have their own migration scripts and ran them to update storybook and the plugins in one go to the latest 6.x.x release. It still works fine locally so let's give this a try in CI.