Fixed a few assets paths issues for monaco-editor
This copies some assets needed on the page level and loads them in the page head if the editor is used on the page.

The monaco css also requires a font and it is expected to live in a very specific relative path so files need to be manually copied into specific folders.

It also looks like a full build was not run before the beta release, so some docs were not auto-generated on other components earlier.

Also I added a new method to allow consumers to change the code content after the component is already loaded.