Fixed a localization fallback issue
Some components use localizable test. The idea implemented was to provide default and allow consumers to override them. The implementation before this PR was to merge a set of defaults with the `resx` prop. However this would only happen once (when the component loaded). So if the parent component would reload and reset the resx value, that new value would replace everything not re-runing the fallback logic.

This PR makes it so localized components now watch the `resx` prop and re-run the fallback logic to store it in a private `localResx` state instead.

Closes #924