Made most input elements form aware
This makes most elements that represent an input be aware of their surrounding form. It currently handles sending the values to the form and reacting to resets. I need to do at least 2 more PRs to complete this:
1. To make dnn-richtext also form aware (need to fix a bug before that)
2. To iron out validation UX/logic when validating a whole form.

I though making those is separate PRs might make it easier to review.