Added examples for usage of files/pictures in forms
Part of #992

We discussed today about making a dnn-file-input and dnn-image-input but then we wanted to make it flexible enough to allow consumers to replace the input with either a progress or a status. After digging into it, we realized that actually the dropzone is already form aware and that a consumer simply has to wrap it into a dnn-fieldset and do his own logic for progress or status hiding the input as needed.

So this PR just adds such usage example to our dnn-example-form and fixes a css position issue for dnn-fieldset to avoid a different padding left and right to make sure the input is centered within the border.