Fixed an issue that caused too few large bundles
Stencil tries to intelligently decide which components should be bundled together for lazy-loading. Somehow this "intelligence" got confused and we ended up with fewer and fewer larger files along the commit history. Rading more about this, we can in the config define how we want our stuff bundled knowing which components usually are used together. The goal is to balance between multiple downloads and small downloads.

This PR manually defines bundles we want.
Examples:
1. dnn-monaco-editor has all the component code and all of monaco code in a single bundle (10.7Mb) that can be forever cached even if we do updates (as long as we don't touch that component).
2. dnn-richtext is in a bundle with all of the Jodit code together (1.1Mb)
3. dnn-color-picker is in a bundle (30Kb) has all the components that are always needed for it (buttons, modals, tabs, tab, etc.)
4. dnn-permissions gid (24Kb) has all the components used by it in the same bundle.
5. So on and so forth reducing in size up to our smallest component that can be used on its own, dnn-chevron (1Kb)