Added `focus` and `blur` logic to `dnn-autocomplete`
This is the overall idea we need for #1055 resolution. It adds a `focus` and `blur` method to our input wrapper and forwards the behaviour to the input field we wrap. The input field itself has the logic to add/remove the focused state on the fieldset itself no matter how the focus is gained or lost.