Fixed some issues with css-variables of inputs
We created a way for people to be able to customize styles of input elements with some css-variables, however some of the variables defined were not being passed down to the wrapping fieldset causing confusion.

Also the namin of those variables was a bit different from component to component, this PR also harmonizes that naming for avoiding further confusion. This rename is a slight breaking change, so I'll call attention to it in the release notes.