Fixed an issue where dnn-select was not being properly initialized
When feeded <option> elements in the slot of dnn-select with a pre-selected item (selected = true) like when used on a form to edit something existing instead of doing something new; The underlying initialization logic was not setting the form level value upon initialization. This caused a discrepency as it would display as valid but would prevent form submission because the form was not valid.

Also the form onInvalid event was not being fired making it even more confusing. This part was fixed in the dnn-button logic.